refactor(landcover): the classifier's own work leaves the handler
The last helpers in terra/actions.py were the land-cover product: the feature
matrix the Random Forest was fitted on, the three model paths, the spectra a
class reflects, the vegetation-index series, and the rasters a run writes. They
become five modules under terra/landcover/, and actions.py drops from 3,508
lines to 2,852.

  features.py   the 80 features per pixel, in the order feature_names.joblib
                records, because another order is other measurements to the head
  classify.py   the three paths, all emitting the same classes, and what a
                class map says once one has run
  spectra.py    what each class reflects, and how far that is from a leaf
  series.py     the index series over the area, per date
  raster.py     the overlay, the GeoTIFF and the two diagnostic ramps

reference/ moves with spectra.py, the only module that reads it, so the path it
computes from __file__ stays right.

ONE DEFINITION OF hex_to_rgb. There were two, one here and one in mapbiomas.py,
each beside its own colour table -- which is the shape of the defect
palette.py was written to end, reappearing in the decoder instead of the table.
It is now in palette.py, with both readers importing it.

THREE ALIASES THAT COULD NOT KEEP THEIR MODULE'S NAME. spectra is the parameter
name of library_limit, series and grid are locals in the actions that call
them. Written plainly, `spectra.library_limit(spectra)` raises
UnboundLocalError, which is what the suite reported.

Verification: pytest 525 passed, 8 skipped. ruff check clean, three contracts
kept, gofmt, go vet and go test -race clean across every package. A ping
through the entry point answers.
