+++ sidecar/terra/landcover/reference/soybean_leaf_reference.json
